Travis CI stalls on Power
This patch moves the keep alive to run during the whole Travis CI build
for Power to avoid it being killed for no output for 10 min.

The Travis CI in Power is more susceptible to it due to the fact it is
running with a reduced number of jobs, running with more jobs can lead
to an issue where bazel server is ended abruptly. Bumping to --jobs=2
has given some performance benefit and so far it has not caused the
issue mentioned, with --jobs=3 it has seen at least once so avoiding
this value for now. If the issue starts to show up in the future we can
move back to --jobs=1.

Signed-off-by: Murilo Fossa Vicentini <muvic@linux.ibm.com>